feat(internal-architecture-refactoring): Tasks 9.1-9.5 - app層重複コード解消
- app/game_state.goのGameState構造体定義を削除し、usecase/game_stateへの型エイリアスに変更
- app/inventory_manager.go、app/statistics_manager.go、app/settings.goを削除
- usecase/game_stateにToSaveData/GameStateFromSaveData関数を追加（persistence.go）
- app層はusecase/game_state層への委譲のみを行う設計に変更
- テストファイルを修正して非公開フィールドへの直接アクセスを解消
